# matcher-api/src/main/java/com/pastryvibe/matcher/domain/FlavorTag.java
package com.pastryvibe.matcher.domain;

public enum FlavorTag {
    SWEET,
    SALTY,
    BUTTERY,
    SPICED,
    FRUITY,
    CHOCOLATEY,
    NUTTY,
    VANILLA,
    CARAMEL,
    CITRUS,
    CREAMY,
    CRUNCHY
}