# matcher-api/src/main/java/com/pastryvibe/matcher/domain/PastryId.java
package com.pastryvibe.matcher.domain;

public enum PastryId {
    STROOPWAFEL,
    TOMPOUCE,
    APPELFLAP,
    CROISSANT,
    PAIN_AU_CHOCOLAT,
    CINNAMON_ROLL,
    DONUT,
    ECLAIR,
    MUFFIN,
    SCONE,
    BAGEL,
    CANNOLI,
    STRUDEL,
    BERLINER,
    BANANA_BREAD,
    CHEESE_DANISH,
    PALMIER,
    PASTEL_DE_NATA,
    BAKLAVA,
    CHOUX_BUN
}