# ratings-api/src/main/java/com/pastryvibe/ratings/api/RatingAcceptedResponse.java
package com.pastryvibe.ratings.api;

import java.time.Instant;
import java.util.UUID;

public record RatingAcceptedResponse(
        UUID eventId,
        Instant acceptedAt,
        String message
) {
}