# ratings-api/src/main/java/com/pastryvibe/ratings/domain/FlavorTag.java
package com.pastryvibe.ratings.domain;

public enum FlavorTag {
    SWEET,
    SALTY,
    BUTTERY,
    SPICED,
    FRUITY,
    CHOCOLATEY,
    NUTTY,
    VANILLA,
    CARAMEL,
    CITRUS,
    CREAMY,
    CRUNCHY
}